whitelist valid dynamic translation
